fornt complete

+++ app/Http/Controllers/Api/Category/CourseCategoryController.php
@@ -6,94 +6,51 @@ use App\CourseCategory;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Category\CategoryStoreRequest;
 use App\Http\Resources\Category\CategoryResource;
-use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Auth;
+use Symfony\Component\HttpFoundation\Response;
 
 class CourseCategoryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index(CourseCategory $category)
     {
         return CategoryResource::collection($category->all());
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(CategoryStoreRequest $request)
     {
         $category = new CourseCategory();
         $category->name = $request->name;
+        $category->slug = Str::slug($request->name , "-");
         $category->description = $request->description;
         $category->user_id = Auth::user()->id;
         $category->save();
         return new CategoryResource($category);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function categoryById(CourseCategory $courseCategory)
     {
 
         return new CategoryResource($courseCategory);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(CategoryStoreRequest $request , CourseCategory $category)
     {
         $category->name = $request->name;
+        $category->slug = Str::slug($request->name , "-");
         $category->description = $request->description;
         $category->update();
         return new CategoryResource($category);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(CourseCategory $category)
     {
+        if(count($category->courses) > 0){
+            return response(['message' => 'Category have Many Courses you donot delete Category']);
+        }
         $category->delete();
-        return response()->json(['data' => $category ,'message' => 'Category Deleted']);
+        return response(['message' => 'Category Deleted Successfully'], Response::HTTP_CREATED);
     }
 }
